Phase L1: money as fixed-point Int64 (bigint e4)
Migrates every money column (numeric(20,4) -> bigint) and every Swift
money value (Decimal -> Int64) to fixed-point at scale 4, per
keepo-local-first-plan.md. Int64 arithmetic is bit-identical across
Postgres, SQLite, and Swift, closing the cross-language rounding-drift
risk that was the original objection to a local-first store.
fx_rates.rate_to_eur is unaffected -- it's a conversion factor, not an
amount.

278/278 pgTAP assertions, 72/72 KeepoCore tests, 9/9 KeepoTests,
xcodebuild clean, swiftlint 0 violations. Migration pushed to hosted.

This is the mandatory L1 human-review stop -- the rest of the
local-first rebuild depends on this representation being correct.

+++ App/BalanceHeaderView.swift
@@ -6,7 +6,7 @@ import SwiftUI
 /// table whenever either screen wants a single aggregate figure of its
 /// own; not retrofitted onto them yet since neither has asked for one.
 struct BalanceHeaderView: View {
-    let amount: Decimal?
+    let amount: Int64?
     let currency: CurrencyInfo?
 
     @Environment(\.isPrivacyMode) private var isPrivacyMode
